Make proto util methods ctors and member methods
